# Demo_Tuan7_Generate_backend_from_openAPI_Version2/src/main/java/com/example/product/api/ApiApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.15.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.product.api;

import com.example.product.model.Product;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.lang.Nullable;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-04T15:13:16.575424593+07:00[Asia/Ho_Chi_Minh]", comments = "Generator version: 7.15.0")
@Validated
@Tag(name = "api", description = "the api API")
public interface ApiApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    public static final String PATH_CREATE_PRODUCT = "/api/products";
    /**
     * POST /api/products : Create a new product
     *
     * @param product  (required)
     * @return Product created (status code 201)
     */
    @Operation(
        operationId = "createProduct",
        summary = "Create a new product",
        responses = {
            @ApiResponse(responseCode = "201", description = "Product created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Product.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ApiApi.PATH_CREATE_PRODUCT,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Product> createProduct(
        @Parameter(name = "Product", description = "", required = true) @Valid @RequestBody Product product
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"price\" : 0.8008282, \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"id\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_DELETE_PRODUCT = "/api/products/{id}";
    /**
     * DELETE /api/products/{id} : Delete a product
     *
     * @param id  (required)
     * @return Product deleted (status code 204)
     */
    @Operation(
        operationId = "deleteProduct",
        summary = "Delete a product",
        responses = {
            @ApiResponse(responseCode = "204", description = "Product deleted")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = ApiApi.PATH_DELETE_PRODUCT
    )
    
    default ResponseEntity<Void> deleteProduct(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_GET_ALL_PRODUCTS = "/api/products";
    /**
     * GET /api/products : Get all products
     *
     * @return A list of products (status code 200)
     */
    @Operation(
        operationId = "getAllProducts",
        summary = "Get all products",
        responses = {
            @ApiResponse(responseCode = "200", description = "A list of products", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Product.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ApiApi.PATH_GET_ALL_PRODUCTS,
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Product>> getAllProducts(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"price\" : 0.8008282, \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"id\" }, { \"price\" : 0.8008282, \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"id\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_GET_PRODUCT_BY_ID = "/api/products/{id}";
    /**
     * GET /api/products/{id} : Get product by ID
     *
     * @param id  (required)
     * @return Product found (status code 200)
     *         or Product not found (status code 404)
     */
    @Operation(
        operationId = "getProductById",
        summary = "Get product by ID",
        responses = {
            @ApiResponse(responseCode = "200", description = "Product found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Product.class))
            }),
            @ApiResponse(responseCode = "404", description = "Product not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ApiApi.PATH_GET_PRODUCT_BY_ID,
        produces = { "application/json" }
    )
    
    default ResponseEntity<Product> getProductById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"price\" : 0.8008282, \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"id\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_UPDATE_PRODUCT = "/api/products/{id}";
    /**
     * PUT /api/products/{id} : Update a product
     *
     * @param id  (required)
     * @param product  (required)
     * @return Product updated (status code 200)
     *         or Product not found (status code 404)
     */
    @Operation(
        operationId = "updateProduct",
        summary = "Update a product",
        responses = {
            @ApiResponse(responseCode = "200", description = "Product updated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Product.class))
            }),
            @ApiResponse(responseCode = "404", description = "Product not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = ApiApi.PATH_UPDATE_PRODUCT,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Product> updateProduct(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id,
        @Parameter(name = "Product", description = "", required = true) @Valid @RequestBody Product product
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"price\" : 0.8008282, \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"id\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}