# Demo_Tuan7_Generate_backend_from_openAPI_Version2/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Product API
  version: 1.0.0
servers:
- url: /
paths:
  /api/products:
    get:
      operationId: getAllProducts
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Product"
                type: array
          description: A list of products
      summary: Get all products
      x-accepts:
      - application/json
    post:
      operationId: createProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
          description: Product created
      summary: Create a new product
      x-content-type: application/json
      x-accepts:
      - application/json
  /api/products/{id}:
    delete:
      operationId: deleteProduct
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Product deleted
      summary: Delete a product
      x-accepts:
      - application/json
    get:
      operationId: getProductById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
          description: Product found
        "404":
          description: Product not found
      summary: Get product by ID
      x-accepts:
      - application/json
    put:
      operationId: updateProduct
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
          description: Product updated
        "404":
          description: Product not found
      summary: Update a product
      x-content-type: application/json
      x-accepts:
      - application/json
components:
  schemas:
    Product:
      example:
        price: 0.8008282
        name: name
        description: description
        id: id
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          format: float
          type: number
      type: object